+++ kwap_common/CHANGELOG.md
@@ -2,4 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 0.4.0 (2021-12-31)
+
+
+### ⚠ BREAKING CHANGES
+
+* **common:** add index bounds to array (#36)
+
+### Features
+
+* **common:** add index bounds to array ([#36](https://github.com/clov-coffee/kwap/issues/36)) ([4198930](https://github.com/clov-coffee/kwap/commit/419893098eb526307b22f414f5d4399256f716d2))
+
+## 0.3.0 (2021-12-31)
+
+
+### ⚠ BREAKING CHANGES
+
+* **common:** add insert trait + bound to array (#35)
+
+### Features
+
+* **common:** add insert trait + bound to array ([#35](https://github.com/clov-coffee/kwap/issues/35)) ([5e8f5ee](https://github.com/clov-coffee/kwap/commit/5e8f5ee40db84d45d224ff33af22cd08b5799365))
+
 ### 0.2.6 (2021-12-31)